% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRi.R
\name{LVNLtest}
\alias{LVNLtest}
\title{Linear and curvilinear model comparison}
\usage{
LVNLtest(x)
}
\arguments{
\item{x}{Dataframe consisting of recombination and adaptation, and maybe 0-fold sites count data.}
}
\value{
A list consisting of a data frame with both models' AICs and two objects with both models.
}
\description{
It allows you to calculate HRi curvilinear and linear functions and compares both models.
}
\details{
This is a first release. It might not work with really complicated datasets, as initial parameters for curvilinear modeling still need to be improved.
}
\keyword{Hill-Robertson}
