% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/HRi.R
\name{rhokbPopFly}
\alias{rhokbPopFly}
\title{PopFly population Rho per kb values}
\usage{
rhokbPopFly(population = "RAL", window = "100kb")
}
\arguments{
\item{population}{String: the desired population as encoded in PopFly. The default value is RAL.}

\item{window}{String: the desired window size. Allowed window sizes: 10kb, 50kb, 100kb,}
}
\value{
A data frame with recombination values in rho/kb, as well as the chromosome, start and end positions of each value.
}
\keyword{PopFly}
\keyword{Recombination}
